Begin work on a math library, and celestial bodies
