+++ src/PatchManager.SassyPatching/Attributes/SassyConstantAttribute.cs
@@ -1,0 +1,8 @@
+﻿namespace PatchManager.SassyPatching.Attributes;
+
+[AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
+public class SassyConstantAttribute : Attribute
+{
+    public string ConstantName;
+    public SassyConstantAttribute(string constantName) => ConstantName = constantName;
+}
